requirements.txt
Flask
easyocr
google-generativeai
gunicorn
